+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [1.0.0-alpha.14] - 2022-12-30
+
 ### Added
 
 #### Core
@@ -271,7 +273,8 @@
 - Basic ruleset using the ESLint rules
 - Rules for module import/export
 
-[unreleased]: https://github.com/arslivinski/eslint-config/compare/1.0.0-alpha.13...HEAD
+[unreleased]: https://github.com/arslivinski/eslint-config/compare/1.0.0-alpha.14...HEAD
+[1.0.0-alpha.14]: https://github.com/arslivinski/eslint-config/compare/1.0.0-alpha.13...1.0.0-alpha.14
 [1.0.0-alpha.13]: https://github.com/arslivinski/eslint-config/compare/1.0.0-alpha.12...1.0.0-alpha.13
 [1.0.0-alpha.12]: https://github.com/arslivinski/eslint-config/compare/1.0.0-alpha.11...1.0.0-alpha.12
 [1.0.0-alpha.11]: https://github.com/arslivinski/eslint-config/compare/1.0.0-alpha.10...1.0.0-alpha.11